+++ scripts/output/projects/FSDP_collect.R
@@ -82,6 +82,44 @@ if (dir.exists(hi_datasets_path)) {
     message("The directory storing health impacts datasets wasn't found. Skipping health impacts.")
 }
 
+magicc7_datasets_path <- "/p/projects/magpie/data/FSEC_magicc7Datasets_raw"
+if (dir.exists(magicc7_datasets_path)) {
+
+    magicc7_datasets <- list.files(magicc7_datasets_path)
+    magicc7_versionToUse  <- grep(rev, magicc7_datasets, value = TRUE)
+
+    if (length(magicc7_versionToUse) == 0) {
+
+        message("No corresponding version ID was found within the MAGICC7 datasets. Using the latest available.")
+
+        highestVersionNr <- max(as.numeric(str_extract(magicc7_datasets, "(?<=v)(.*?)(?=_)")))
+        magicc7_versionToUse <- grep(paste0("v", highestVersionNr), magicc7_datasets, value = TRUE)
+
+    } else if (length(magicc7_versionToUse) >= 2) {
+        stop("Duplicated version IDs were found in the MAGICC7 datasets, only one is expected.")
+    }
+
+    magicc7_versionToUse_path <- file.path(magicc7_datasets_path, magicc7_versionToUse)
+
+  .appendMAGICC7 <- function(.x) {
+      cfg <- gms::loadConfig(file.path(.x, "config.yml"))
+      title <- cfg$title
+
+      tryCatch(
+          expr = {
+              appendReportMAGICC7(resultsPath = magicc7_versionToUse_path, scenario = title, dir = .x)
+          }, error = function(e) {
+              message("Unable to append MAGICC7 dataset for scenario: ", title)
+          }
+      )
+  }
+  lapply(X = outputdir, FUN = .appendMAGICC7)
+
+} else {
+    message("The directory storing MAGICC7 datasets wasn't found. Skipping AR6 global warming calculations.")
+}
+
+
 ##########
 # Generate output files
 cat("\nStarting output generation\n")
@@ -265,15 +303,23 @@ var_reg <- c(indicators_main,
              "Income|Number of People Below 1p90 USDppp11/day",
              "Income|Number of People Below 3p20 USDppp11/day",
              "Income|Number of People Below 5p50 USDppp11/day",
-             "Health|Attributable deaths|Risk|Diet and anthropometrics",
-             "Health|Percent change in Attributable deaths|Risk|Diet and anthropometrics",
-             "Health|Years of life lost|Risk|Diet and anthropometric",
-             "Health|Percent change in Years of life lost|Risk|Diet and anthropometrics"
+             "Health|Years of life lost|Disease",
+             "Health|Years of life lost|Disease|+|Congenital Heart Disease",
+             "Health|Years of life lost|Disease|+|Stroke",
+             "Health|Years of life lost|Disease|+|Cancer",
+             "Health|Years of life lost|Disease|+|Type-2 Diabetes",
+             "Health|Years of life lost|Disease|+|Respiratory Disease",
+             "Health|Percent change in Years of life lost|Disease",
+             "Health|Percent change in Years of life lost|Disease|+|Congenital Heart Disease",
+             "Health|Percent change in Years of life lost|Disease|+|Stroke",
+             "Health|Percent change in Years of life lost|Disease|+|Cancer",
+             "Health|Percent change in Years of life lost|Disease|+|Type-2 Diabetes",
+             "Health|Percent change in Years of life lost|Disease|+|Respiratory Disease"
              )
 var_reg <- unique(var_reg)
 
 var_iso <- c("Population",
-             "Health|Years of life lost|Risk|Diet and anthropometrics",
+             "Health|Years of life lost|Disease",
              "Labor|Employment|Agricultural employment",
              "Nutrition|Anthropometrics|People underweight",
              "Nutrition|Anthropometrics|People obese",
@@ -325,8 +371,13 @@ for (i in 1:length(outputdir)) {
     } else missing <- c(missing,nc_file)
 
     ## Gridded temperature data from ISIMIP archive for relevant SSP/RCP
-    rcp     <- if (thisScen == "BAU") "ssp245" else "ssp119"
-    nc_file <- "./input/FSEC_GlobalSurfaceTempPerRCP_v2_16-01-23/FSEC_GlobalSurfaceTempPerRCP_v2_16-01-23.mz"
+    rcp <- switch(thisScen,
+         "BAU"      = "ssp460",
+         "FSDP"     = "ssp119",
+         "SSP2fsdp" = "ssp245",
+         "Invalid case")
+
+    nc_file <- "./input/FSEC_GlobalSurfaceTempPerRCP_v3_04-05-23/FSEC_GlobalSurfaceTempPerRCP_v3_04-05-23.mz"
     if (file.exists(nc_file)) {
       a <- read.magpie(nc_file)[, years, rcp]
       getNames(a) <- "Global Surface Temperature (C)"
@@ -388,17 +439,35 @@ for (i in 1:length(outputdir)) {
       getSets(a,fulldim = F)[3] <- "variable"
       a <- addLocation(a)
       y <- mbind(y,a)
-    } else missing <- c(missing,nc_file)
+    } else missing <- c(missing, nc_file)
 
     ## Water
     nc_file <- file.path(outputdir[i], "watStressViolations.mz")
     if (file.exists(nc_file)) {
       a <- read.magpie(nc_file)[, years, ]
       getNames(a) <- "water stress and violations"
-      getSets(a,  fulldim = FALSE)[3] <- "variable"
+      getSets(a, fulldim = FALSE)[3] <- "variable"
       a <- addLocation(a)
       y <- mbind(y, a)
-    } else missing <- c(missing,nc_file)
+    } else missing <- c(missing, nc_file)
+
+    nc_file <- file.path(outputdir[i], "efvVolume.mz")
+    if (file.exists(nc_file)) {
+      a <- read.magpie(nc_file)[, years, ]
+      getNames(a) <- "water environmental flow violations volume (km3)"
+      getSets(a, fulldim = FALSE)[3] <- "variable"
+      a <- addLocation(a)
+      y <- mbind(y, a)
+    } else missing <- c(missing, nc_file)
+
+    nc_file <- file.path(outputdir[i], "efvVolume_ha.mz")
+    if (file.exists(nc_file)) {
+      a <- read.magpie(nc_file)[, years, ]
+      getNames(a) <- "water environmental flow violations volume (m3/ha)"
+      getSets(a, fulldim = FALSE)[3] <- "variable"
+      a <- addLocation(a)
+      y <- mbind(y, a)
+    } else missing <- c(missing, nc_file)
 
     #add dimensions