+++ CHANGELOG.md
@@ -9,7 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [4.6.6] - 2023-05-05
 
 ### changed
-- finalize FSEC scripts (new indicators and scenarios)
+- **config** updated scenario_fsec.csv to reflect new GST validation
+- **scripts** finalize FSEC scripts (new indicators and scenarios)
+- **scripts** included new output indicator for water
+- **scripts** updated global surface temperature maps to new RCPs per scenario
 
 
 ## [4.6.5] - 2023-03-29